+++ index.html
@@ -20,7 +20,38 @@
       <ul class="comments">
       </ul>
         <p class="download">Жди, жди, у меня тут куча информации, ее надо обработать!</p>
-        <div class="add-form remove-div">
+      <div class="add-form remove-div remove-for-login">
+        <input
+          type="text"
+          class="add-form-name"
+          placeholder="Введите ваше имя"
+        />
+        <textarea
+          type="textarea"
+          class="add-form-text"
+          placeholder="Введите ваш коментарий"
+          rows="4"
+        ></textarea>
+        <div class="add-form-row">
+          <button class="add-form-button">Написать</button>
+        </div>
+      </div>
+      <div class="add-form form_login remove-div">
+        <input
+          type="text"
+          class="add-form-name form_login_login"
+          placeholder="Логин"
+        />
+        <textarea
+          type="textarea"
+          class="add-form-text form_login_password"
+          placeholder="Введите ваш Пароль"
+          rows="4"
+        ></textarea>
+        <div class="add-form-row">
+          <button class="add-form-button add-login-button">Авторизоваться</button>
+        </div>
+       <p>Нет аккаунта?</p><a href="/">Зарегистрируйтесь</a>
       </div>
     </div>
     <script type="module" src="main.js"></script>
